refactor(repos): made mapper a TS, and search util - mapper
